+++ code/src/cljs/sixsq/nuvla/ui/acl/views.cljs
@@ -340,9 +340,9 @@
 
 
 (defn AclButton
-  [{:keys [default-value read-only] :as opts}]
+  [{:keys [default-value read-only default-active?] :as opts}]
   (let [tr      (subscribe [::i18n-subs/tr])
-        active? (r/atom false)
+        active? (r/atom default-active?)
         acl     (or default-value (when-not read-only {:owners [@(subscribe [::authn-subs/user-id])]}))
         ui-acl  (when acl (r/atom (utils/acl->ui-acl-format acl)))]
     (fn [opts]