+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/events.cljs
@@ -6,6 +6,8 @@
     [sixsq.nuvla.ui.edge.effects :as edge-fx]
     [sixsq.nuvla.ui.edge.events :as edge-events]
     [sixsq.nuvla.ui.history.events :as history-events]
+    [sixsq.nuvla.ui.messages.events :as messages-events]
+    [sixsq.nuvla.ui.utils.response :as response]
     [taoensso.timbre :as log]))
 
 
@@ -37,8 +39,8 @@
     (cond-> {::cimi-api-fx/get    [id #(dispatch [::set-nuvlabox %])
                                    :on-error #(dispatch [::set-nuvlabox nil])]
              ::cimi-api-fx/search [:nuvlabox-peripheral
-                                   {:filter (str "parent='" id "'")
-                                    :last   10000
+                                   {:filter  (str "parent='" id "'")
+                                    :last    10000
                                     :orderby "id"}
                                    #(dispatch [::set-nuvlabox-peripherals %])]}
             (not= (:id nuvlabox) id) (assoc :db (merge db spec/defaults)))))
@@ -50,6 +52,19 @@
     (let [nuvlabox-id (:id nuvlabox)]
       {::cimi-api-fx/operation [nuvlabox-id "decommission"
                                 #(dispatch [::get-nuvlabox nuvlabox-id])]})))
+
+(reg-event-fx
+  ::edit
+  (fn [_ [_ resource-id data]]
+    {::cimi-api-fx/edit [resource-id data
+                         #(if (instance? js/Error %)
+                            (let [{:keys [status message]} (response/parse-ex-info %)]
+                              (dispatch [::messages-events/add
+                                         {:header  (cond-> (str "error editing " resource-id)
+                                                           status (str " (" status ")"))
+                                          :content message
+                                          :type    :error}]))
+                            (dispatch [::set-nuvlabox %]))]}))
 
 
 (reg-event-fx